Predictors of Spring Wildfires in the UK

This RasterStack object contains a collection of raster layers representing various environmental predictors related to spring wildfires in the UK.

spring_independent_var_stack

Format

A RasterStack object with the following layers:

Slope

Raster layer representing the slope of the terrain.

Aspect

Raster layer representing the aspect of the terrain.

Average.Temperature

Raster layer representing the average temperature during spring.

Precipitation

Raster layer representing precipitation during spring.

Wind.Speed

Raster layer representing average wind speed during spring.

Proximity.to.Major.Roads

Raster layer representing proximity to major roads.

Population.Counts

Raster layer representing population counts in the UK.

Details

The RasterStack has the following properties:

class: RasterStack

dimensions: 263 rows, 250 columns, 65750 cells, 7 layers

resolution: 0.04166667 x 0.04166667 (x, y)

extent: -8.666667, 1.75, 49.875, 60.83333 (xmin, xmax, ymin, ymax)

crs: +proj=longlat +datum=WGS84 +no_defs

names: Slope, Aspect, Average.Temperature, Precipitation, Wind.Speed, Proximity.to.Major.Roads, Population.Counts

min values: 0.00000, 0.00000, 0.00000, 33.96667, 2.93600, -0.09200, 0.00000

max values: 0.0930622, 6.2831853, 10.5779605, 193.2136383, 10.3826666, 68.3430023, 441.3430481

See also

Author

Matteo Larrode